+++ src/tests/functional/plugin/gna/Import_export_tests/base/import_export_base.cpp
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 //
 
-#include "base/import_export_base/import_export_base.hpp"
+#include "import_export_base.hpp"
 
 #include <fstream>
