+++ src/tests/functional/plugin/gna/Import_export_tests/import_export_multi_inputs.cpp
@@ -10,7 +10,7 @@
 #include <fstream>
 
 #include "ngraph_functions/builders.hpp"
-#include "base/import_export_base/import_export_base.hpp"
+#include "base/import_export_base.hpp"
 
 namespace LayerTestsDefinitions {
