--- src/tests/functional/plugin/shared/include/import_export_tests/import_reshape_permute_conv.hpp
@@ -1,16 +1,0 @@
-// Copyright (C) 2018-2022 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-//
-
-#pragma once
-
-#include "base/import_export_base/import_export_base.hpp"
-
-namespace LayerTestsDefinitions {
-
-class ImportReshapePermuteConv : public FuncTestUtils::ImportNetworkTestBase {
-protected:
-    void SetUp() override;
-};
-
-} // namespace LayerTestsDefinitions
